// packages/zibri/src/utilities/metadata-injection-keys.enum.ts
/**
 * The keys used for defining metadata.
 */
export enum MetadataInjectionKeys {
    FILE_LOCATION = 'file:location',
    // eslint-disable-next-line cspell/spellchecker
    PARAM_TYPES = 'design:paramtypes',
    TYPE = 'design:type',
    DI_TOKEN = 'di:token',
    DI_INJECT_PARAM_TOKENS = 'di:inject_param_tokens',
    DI_INJECT_PARAM_OPTIONS = 'di:inject_param_options',
    CONTROLLER_DATA = 'controller:data',
    CONTROLLER_ROUTES = 'controller:routes',
    CONTROLLER_IS_LOGGED_IN = 'controller:is_logged_in',
    CONTROLLER_SKIP_IS_LOGGED_IN = 'controller:skip_is_logged_in',
    CONTROLLER_IS_NOT_LOGGED_IN = 'controller:is_not_logged_in',
    CONTROLLER_SKIP_IS_NOT_LOGGED_IN = 'controller:skip_is_not_logged_in',
    CONTROLLER_HAS_ROLE = 'controller:has_role',
    CONTROLLER_SKIP_HAS_ROLE = 'controller:skip_has_role',
    CONTROLLER_BELONGS_TO = 'controller:belongs_to',
    CONTROLLER_SKIP_BELONGS_TO = 'controller:skip_belongs_to',
    CONTROLLER_REQUIRE_2FA = 'controller:require_2fa',
    CONTROLLER_SKIP_REQUIRE_2FA = 'controller:skip_require_2fa',
    ROUTE_PATH_PARAMS = 'route:path_params',
    ROUTE_QUERY_PARAMS = 'route:query_params',
    ROUTE_HEADER_PARAMS = 'route:header_params',
    ROUTE_BODY = 'route:body',
    ROUTE_RESPONSES = 'route:responses',
    ROUTE_CURRENT_USER = 'route:current_user',
    ROUTE_IS_LOGGED_IN = 'route:is_logged_in',
    ROUTE_SKIP_IS_LOGGED_IN = 'route:skip_is_logged_in',
    ROUTE_IS_NOT_LOGGED_IN = 'route:is_not_logged_in',
    ROUTE_SKIP_IS_NOT_LOGGED_IN = 'route:skip_is_not_logged_in',
    ROUTE_SKIP_HAS_ROLE = 'route:skip_has_role',
    ROUTE_HAS_ROLE = 'route:has_role',
    ROUTE_BELONGS_TO = 'route:belongs_to',
    ROUTE_SKIP_BELONGS_TO = 'route:skip_belongs_to',
    ROUTE_REQUIRE_2FA = 'route:require_2fa',
    ROUTE_SKIP_REQUIRE_2FA = 'route:skip_require_2fa',
    ROUTE_SKIP_AUTH = 'route:skip_auth',
    MODEL_PROPERTIES = 'model:properties',
    MODEL_RELATIONS = 'model:relations',
    ENTITY_METADATA = 'entity:metadata',
    BACKUP_RESOURCE_METADATA = 'backup_resource:metadata',
    WEBSOCKET_CONTROLLER_DATA = 'websocket_controller:data',
    WEBSOCKET_CONTROLLER_ROUTES = 'websocket_controller:routes',
    ROUTE_CURRENT_WEBSOCKET_CONNECTION = 'route:current_websocket_connection',
    EXCLUDED_PROPERTY_VALUE = 'entity:excluded_property_value'
}